+++ scrapli_community/cumulus/vtysh/sync_driver.py
@@ -1,0 +1,37 @@
+"""scrapli_community.cumulus.vtysh.sync_driver"""
+from scrapli.driver import NetworkDriver
+
+
+def default_sync_on_open(conn: NetworkDriver) -> None:
+    """
+    cumulus_vtysh on_open callable
+
+    Args:
+        conn: NetworkDriver object
+
+    Returns:
+        None
+
+    Raises:
+        N/A
+    """
+    conn.acquire_priv(desired_priv=conn.default_desired_privilege_level)
+
+
+def default_sync_on_close(conn: NetworkDriver) -> None:
+    """
+    cumulus_vtysh default on_close callable
+
+    Args:
+        conn: NetworkDriver object
+
+    Returns:
+        None
+
+    Raises:
+        N/A
+
+    """
+    conn.acquire_priv(desired_priv=conn.default_desired_privilege_level)
+    conn.channel.write(channel_input="exit")
+    conn.channel.send_return()